% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsicore_utils.R
\name{clamp01}
\alias{clamp01}
\title{Clamp values to the unit interval}
\usage{
clamp01(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Clamped values
}
\description{
Clamp values to the unit interval
}
