% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{compute_data_hash}
\alias{compute_data_hash}
\title{Compute data hash for caching}
\usage{
compute_data_hash(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Character hash
}
\description{
Compute data hash for caching
}
