% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsicore_scoring.R
\name{compute_dsi}
\alias{compute_dsi}
\title{Compute DSI score (with outlier penalty)}
\usage{
compute_dsi(metrics, weights = default_dsi_weights())
}
\arguments{
\item{metrics}{Named list of metrics}

\item{weights}{DSI weights}
}
\value{
DSI score (0-100)
}
\description{
Compute DSI score (with outlier penalty)
}
