% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsicore_metrics.R
\name{default_dsi_refs}
\alias{default_dsi_refs}
\title{Default reference values for DSI computation}
\usage{
default_dsi_refs()
}
\value{
List of reference parameters
}
\description{
Default reference values for DSI computation
}
