% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsicore_selection.R
\name{default_selection_opts}
\alias{default_selection_opts}
\title{Default selection options}
\usage{
default_selection_opts()
}
\value{
List of selection thresholds and criteria
}
\description{
Default selection options
}
