% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsicore_utils.R
\name{.append_reason}
\alias{.append_reason}
\title{Append reason code}
\usage{
.append_reason(reasons, reason)
}
\arguments{
\item{reasons}{Current reason string}

\item{reason}{New reason to append}
}
\value{
Combined reason string
}
\description{
Append reason code
}
