% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsicore_utils.R
\name{.psi_numeric}
\alias{.psi_numeric}
\title{Population Stability Index}
\usage{
.psi_numeric(x_ref, x_new, n_bins = 10, eps = 0.005)
}
\arguments{
\item{x_ref}{Reference sample}

\item{x_new}{New sample}

\item{n_bins}{Number of bins}

\item{eps}{Smoothing constant}
}
\value{
PSI value
}
\description{
Population Stability Index
}
