% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsicore_utils.R
\name{.safe_ks_pvalue}
\alias{.safe_ks_pvalue}
\title{Safe KS test p-value}
\usage{
.safe_ks_pvalue(x_ref, x_new)
}
\arguments{
\item{x_ref}{Reference sample}

\item{x_new}{New sample}
}
\value{
KS test p-value or NA
}
\description{
Safe KS test p-value
}
