% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsicore_utils.R
\name{.safe_spearman}
\alias{.safe_spearman}
\title{Safe Spearman correlation}
\usage{
.safe_spearman(x, y)
}
\arguments{
\item{x, y}{Numeric vectors}
}
\value{
Spearman correlation or NA
}
\description{
Safe Spearman correlation
}
