% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsicore_utils.R
\name{.year_in_window}
\alias{.year_in_window}
\title{Check if year is in window}
\usage{
.year_in_window(year, start_year, end_year)
}
\arguments{
\item{year}{Year vector}

\item{start_year}{Window start year}

\item{end_year}{Window end year}
}
\value{
Logical vector
}
\description{
Check if year is in window
}
