% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_theme.R
\name{dsi_card}
\alias{dsi_card}
\title{Card without default chrome}
\usage{
dsi_card(..., title = NULL, class = NULL, actions = NULL, id = NULL)
}
\description{
Card without default chrome
}
\keyword{internal}
