% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_state.R
\name{dsi_step_unlocked}
\alias{dsi_step_unlocked}
\title{Which steps can be opened, given what has been completed}
\usage{
dsi_step_unlocked(step, state)
}
\description{
Which steps can be opened, given what has been completed
}
\keyword{internal}
