% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explore.R
\name{dsi_v2_limiters}
\alias{dsi_v2_limiters}
\title{Explain which DSI_v2 multipliers pull a window's score down}
\usage{
dsi_v2_limiters(m, threshold = 0.9)
}
\description{
Explain which DSI_v2 multipliers pull a window's score down
}
\keyword{internal}
