% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{dsi_write_report}
\alias{dsi_write_report}
\title{Write the HTML report for the current state}
\usage{
dsi_write_report(app_state, file)
}
\description{
Write the HTML report for the current state
}
\keyword{internal}
