% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsiStudio-package.R
\docType{package}
\name{dsiStudio-package}
\alias{dsiStudio}
\alias{dsiStudio-package}
\title{dsiStudio: DSI Studio - Data Suitability Index Screening Tool}
\description{
Interactive Shiny application for screening fisheries catch/effort data using the Data Suitability Index (DSI). Scores time windows for suitability in Fox/Schaefer surplus production models.
}
\author{
\strong{Maintainer}: Javier Porobic \email{jporobicg@gmail.com}

Authors:
\itemize{
  \item Javier Porobic \email{jporobicg@gmail.com}
}

}
\keyword{internal}
