% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_theme.R
\name{format_dsi_score}
\alias{format_dsi_score}
\title{Format DSI score for display}
\usage{
format_dsi_score(score, digits = 0)
}
\arguments{
\item{score}{Numeric DSI score(s)}

\item{digits}{Number of decimal places}
}
\description{
Format DSI score for display
}
