% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsicore_selection.R
\name{get_dsi_band}
\alias{get_dsi_band}
\title{Get DSI band label and color}
\usage{
get_dsi_band(score, cutpoints = c(poor = 0, moderate = 50, good = 70))
}
\arguments{
\item{score}{DSI or DSI_v2 score}

\item{cutpoints}{Named vector of cutpoints}
}
\value{
List with label and color
}
\description{
Get DSI band label and color
}
