% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{load_demo_data}
\alias{load_demo_data}
\title{Load demo dataset}
\usage{
load_demo_data(dataset_name = c("main_groups", "species_fleet"))
}
\arguments{
\item{dataset_name}{Name of demo dataset ("main_groups" or "species_fleet")}
}
\value{
Data frame (tibble)
}
\description{
Reads one of the example datasets shipped in \code{inst/demo_data}.
}
