% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsicore_data.R
\name{make_group_key}
\alias{make_group_key}
\title{Create group key from grouping columns}
\usage{
make_group_key(df, group_cols = c("species", "fleet"))
}
\arguments{
\item{df}{Data frame}

\item{group_cols}{Column names to group by}
}
\value{
Character vector of group keys
}
\description{
Create group key from grouping columns
}
