% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_audit.R
\name{mod_audit_server}
\alias{mod_audit_server}
\title{Audit server}
\usage{
mod_audit_server(id, app_state)
}
\arguments{
\item{id}{Module namespace id}

\item{app_state}{Shared \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} holding the app state}
}
\description{
Audit server
}
