% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare.R
\name{mod_compare_server}
\alias{mod_compare_server}
\title{Comparison server}
\usage{
mod_compare_server(id, app_state)
}
\arguments{
\item{id}{Module namespace id}

\item{app_state}{Shared \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} holding the app state}
}
\description{
Comparison server
}
