% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_context_rail.R
\name{mod_context_rail_server}
\alias{mod_context_rail_server}
\title{Context rail server}
\usage{
mod_context_rail_server(id, app_state)
}
\arguments{
\item{id}{Module namespace id}

\item{app_state}{Shared \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} holding the app state}
}
\description{
Context rail server
}
