% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_input.R
\name{mod_data_input_server}
\alias{mod_data_input_server}
\title{Data input server}
\usage{
mod_data_input_server(id, app_state)
}
\arguments{
\item{id}{Module namespace id}

\item{app_state}{Shared \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} holding the app state}
}
\description{
Data input server
}
