% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_decide.R
\name{mod_decide_server}
\alias{mod_decide_server}
\title{Decide server}
\usage{
mod_decide_server(id, app_state)
}
\arguments{
\item{id}{Module namespace id}

\item{app_state}{Shared \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} holding the app state}
}
\description{
Decide server
}
