% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explore.R
\name{mod_explore_server}
\alias{mod_explore_server}
\title{Explore server}
\usage{
mod_explore_server(id, app_state)
}
\arguments{
\item{id}{Module namespace id}

\item{app_state}{Shared \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} holding the app state}
}
\description{
Explore server
}
