% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report_server}
\alias{mod_report_server}
\title{Report server}
\usage{
mod_report_server(id, app_state)
}
\arguments{
\item{id}{Module namespace id}

\item{app_state}{Shared \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} holding the app state}
}
\description{
Report server
}
