% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_screen.R
\name{mod_screen_server}
\alias{mod_screen_server}
\title{Screen server}
\usage{
mod_screen_server(id, app_state)
}
\arguments{
\item{id}{Module namespace id}

\item{app_state}{Shared \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} holding the app state}
}
\description{
Screen server
}
