% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_theme.R
\name{next_step_button}
\alias{next_step_button}
\title{"Continue" button that navigates to the next step}
\usage{
next_step_button(step, label)
}
\description{
"Continue" button that navigates to the next step
}
\keyword{internal}
