% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsicore_utils.R
\name{safe_log_cpue}
\alias{safe_log_cpue}
\title{Safe log transformation of CPUE}
\usage{
safe_log_cpue(cpue)
}
\arguments{
\item{cpue}{Numeric vector of CPUE values}
}
\value{
List with log_cpue, n_invalid, invalid_idx
}
\description{
Safe log transformation of CPUE
}
