% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explore.R
\name{window_eligibility}
\alias{window_eligibility}
\title{Why a single (possibly user-chosen) window is or is not eligible}
\usage{
window_eligibility(m, opts = default_selection_opts())
}
\description{
Why a single (possibly user-chosen) window is or is not eligible
}
\keyword{internal}
