% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{guess_column_mapping}
\alias{guess_column_mapping}
\title{Guess column mapping from data frame}
\usage{
guess_column_mapping(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
List of column mappings
}
\description{
Guess column mapping from data frame
}
